+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - UI hiding when actively using seek or volume slider
 - Empty background boxes with TTML subtitles on Chromecast
 - Default selection not highlighted in AudioTrackListBox and SubtitleListBox
+- Clear subtitles list when source is unloaded
 
 ## [3.9.2]